Refactor a little, prepare for implementing ENCRYPT2
