+++ src/AdoNetCore.AseClient/Packet/LoginPacket.cs
@@ -36,6 +36,13 @@ namespace AdoNetCore.AseClient.Packet
         public int PacketSize { get; set; }
         public bool EncryptPassword { get; set; }
 
+        public static void ResetPreferredLSecLogin()
+        {
+            PreferredLSecLogin = LSecLogin.TDS_SEC_LOG_ENCRYPT3;
+        }
+
+        public static LSecLogin PreferredLSecLogin { get; set; } = LSecLogin.TDS_SEC_LOG_ENCRYPT3;
+
         public LoginPacket(string hostname, string username, string password, string processId, string applicationName, string serverName, string language, string charset, string clientLibrary, int packetSize, CapabilityToken capability, bool encryptPassword)
         {
             Capability = capability;
@@ -106,9 +113,9 @@ namespace AdoNetCore.AseClient.Packet
             stream.WritePaddedString(Language, TDS_MAXNAME, env.Encoding); //llanguage
 
             var lSecLogin = EncryptPassword
-                //? LSecLogin.TDS_SEC_LOG_ENCRYPT | LSecLogin.TDS_SEC_LOG_ENCRYPT2 | LSecLogin.TDS_SEC_LOG_ENCRYPT3
-                ? LSecLogin.TDS_SEC_LOG_ENCRYPT3
+                ? PreferredLSecLogin
                 : LSecLogin.UNUSED;
+
             stream.Write(new byte[]
             {
                 (byte)LSetLang,
